moved around redux store for cleanup
