+++ src/Admin/store/admin/actions.js
@@ -4,13 +4,13 @@ import {
     ADMIN_LOGIN_USER,
     ADMIN_LOGIN_USER_SUCCESS,
     ADMIN_LOGIN_USER_ERROR
-} from '../actions';
-import { URL, KEY } from '../../constants/defaultValues';
+} from '../../../redux/actions.js';
+import { URL, KEY } from '../../../constants/defaultValues.js';
 import {
     setCurrentUser,
     getNormalHeaders,
     openNotificationWithIcon
-} from '../../helpers/Utils';
+} from '../../../helpers/Utils.js';
 
 export const adminLoginUserSuccess = (user) => async (dispatch) => {
     dispatch({
