+++ src/Admin/store/adminEvalutors/reducer.js
@@ -1,33 +1,31 @@
 // Foulders Reducers //
 import {
-    ADMIN_LOGIN_USER,
-    ADMIN_LOGIN_USER_SUCCESS,
-    ADMIN_LOGIN_USER_ERROR,
-} from '../actions';
+    ADMIN_EVALUTORS_LIST,
+    ADMIN_EVALUTORS_LIST_SUCCESS,
+    ADMIN_EVALUTORS_LIST_ERROR,
+} from '../../../redux/actions.js';
 
 const INIT_STATE = {
-    currentUser: {},
-    loading: false,
-    error: '',
+    evalutorsList: [],
 };
 
 export default (state = INIT_STATE, action) => {
     const newState = { ...state };
     switch (action.type) {
-    case ADMIN_LOGIN_USER:
+    case ADMIN_EVALUTORS_LIST:
         return { ...state, loading: true, error: '' };
-    case ADMIN_LOGIN_USER_SUCCESS:
+    case ADMIN_EVALUTORS_LIST_SUCCESS:
         return {
             ...state,
             loading: false,
-            currentUser: action.payload,
+            evalutorsList: action.payload,
             error: '',
         };
-    case ADMIN_LOGIN_USER_ERROR:
+    case ADMIN_EVALUTORS_LIST_ERROR:
         return {
             ...state,
             loading: false,
-            currentUser: null,
+            evalutorsList: [],
             error: action.payload.message,
         };
     default:
