+++ src/Teachers/store/teacher/reducer.js
@@ -1,31 +1,33 @@
 // Foulders Reducers //
 import {
-    ADMIN_EVALUTORS_LIST,
-    ADMIN_EVALUTORS_LIST_SUCCESS,
-    ADMIN_EVALUTORS_LIST_ERROR,
-} from '../actions';
+    TEACHER_LOGIN_USER,
+    TEACHER_LOGIN_USER_SUCCESS,
+    TEACHER_LOGIN_USER_ERROR,
+} from '../../../redux/actions.js';
 
 const INIT_STATE = {
-    evalutorsList: [],
+    currentUser: {},
+    loading: false,
+    error: '',
 };
 
 export default (state = INIT_STATE, action) => {
     const newState = { ...state };
     switch (action.type) {
-    case ADMIN_EVALUTORS_LIST:
+    case TEACHER_LOGIN_USER:
         return { ...state, loading: true, error: '' };
-    case ADMIN_EVALUTORS_LIST_SUCCESS:
+    case TEACHER_LOGIN_USER_SUCCESS:
         return {
             ...state,
             loading: false,
-            evalutorsList: action.payload,
+            currentUser: action.payload,
             error: '',
         };
-    case ADMIN_EVALUTORS_LIST_ERROR:
+    case TEACHER_LOGIN_USER_ERROR:
         return {
             ...state,
             loading: false,
-            evalutorsList: [],
+            currentUser: null,
             error: action.payload.message,
         };
     default:
