+++ src/redux/reducers.js
@@ -1,14 +1,14 @@
 import { combineReducers } from 'redux';
 import authUser from './auth/reducer';
-import admin from './admin/reducer';
-import teacher from './teacher/reducer';
+import admin from '../Admin/store/admin/reducer';
+import teacher from '../Teachers/store/teacher/reducer';
 import sample from './sample/reducers';
-import mentors from './mentors/reducer';
+import mentors from '../Teachers/store/mentors/reducer';
 import modules from './modules/reducer';
-import adminCourses from './adminCourses/reducer';
-import adminEvalutors from './adminEvalutors/reducer';
-import adminMentors from './adminMentors/reducer';
-import adminNotifications from './adminNotifications/reducer';
+import adminCourses from '../Admin/Courses/store/adminCourses/reducer';
+import adminEvalutors from '../Admin/store/adminEvalutors/reducer';
+import adminMentors from '../Admin/store/adminMentors/reducer';
+import adminNotifications from '../Admin/store/adminNotifications/reducer';
 import schoolRegistration from './schoolRegistration/reducers';
 import evaluatorsBulkUpload from './evaluatorsBulkUpload/reducers';
